corrected the fprintf arguments
git-svn-id: https://qthreads.googlecode.com/svn/trunk@575 9dbd018f-d1da-d342-4c11-58c73dfe904b

+++ test/test2.c
@@ -64,7 +64,7 @@ int main(int argc, char *argv[])
     qthread_finalize();
 
     if (interactive == 1) {
-	fprintf(stderr, "Final value of x=%d\n", x);
+	fprintf(stderr, "Final value of x=%lu\n", (unsigned long)x);
     }
 
     if (x == target)
